eod/requirements.txt
DateTime==4.3
Django==2.2.1
django-admin==1.3.2
django-cors-headers==2.5.3
django-excel-response2==2.0.8
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.9.3
djangorestframework-expiring-authtoken==0.1.4
gunicorn==19.9.0
psycopg2==2.8.2
pytz==2019.1
screen==1.0.1
sqlparse==0.3.0
whitenoise==4.1.2
xlwt==1.3.0
zope.interface==4.6.0
